requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
PyMuPDF==1.24.5
python-docx==1.1.2
scikit-learn==1.5.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
numpy==1.26.4
